Fix release content
From karl@freefriends.org

1) The chklref.tds.zip file is not necessary (and also not correct --
the chklref.tds/ top level is extraneous). I recommend omitting it in
the future. Alternatively, if it is included, in the upload to CTAN it
should be at the "same level" as the unpacked directory (not inside it),
as shown at https://ctan.org/help/upload-pkg.